+++ coffeeCompass/src/main/java/cz/fungisoft/coffeecompass/repository/ImageRepository.java
@@ -1,11 +1,8 @@
 package cz.fungisoft.coffeecompass.repository;
 
-import java.util.List;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import cz.fungisoft.coffeecompass.entity.Comment;
 import cz.fungisoft.coffeecompass.entity.Image;
 
 public interface ImageRepository extends JpaRepository<Image, Integer>
@@ -14,5 +11,5 @@ public interface ImageRepository extends JpaRepository<Image, Integer>
     public Integer getSiteIdForImage(Integer commentId);
     
     @Query("select im from Image im WHERE coffeeSite.id=?1")
-    public Image getImageForSite(Integer coffeeSiteID);
+    public Image getImageForSite(Long coffeeSiteID);
 }
